fix[neovim]: arrow key not (dis/en)able correctly
When enter o neotree pop-up arrow left and right arrow keys are need to
better experience but I disable this by default on any mode other than
cmdline mode.

So this need to be enable on neotree buffer but previously this was
buggy because when cancel pop-up some event that a relay on isn't called
so moved all logic to buffer enter/leave events fix this.